// src/__generated__/ExerciseVariables.graphql.ts
export type ExerciseVariables$data = {
  readonly exerciseVariables: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly createdAt: string;
        readonly description: string | null | undefined;
        readonly id: string;
        readonly name: string;
        readonly unit: string | null | undefined;
        readonly updatedAt: string;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null | undefined;
    };
  };
  readonly " $fragmentType": "ExerciseVariables";
};
export type ExerciseVariables$key = {
  readonly " $data"?: ExerciseVariables$data;
  readonly " $fragmentSpreads": FragmentRefs<"ExerciseVariables">;
};